Moved test_helper to own repo/project
Cloned shr2js repo and stripped out everything but the test_helper (using technique described here: https://imwill.com/split-a-large-git-repository/).

Moved 'test/test_helper' to 'common-export-tests' to more properly reflect what it is -- and to move it out of 'test' folder so it can be imported by other modules.

As a result of trimming the shr2js repo, commit messages before this commit may not always make sense in the context of this repository (since they may reference changes in importer or exporters).